requirements.txt
h5py==2.10.0
imageio==2.6.1
matplotlib==3.1.3
numpy==1.18.1
opencv-python==4.2.0.32
pytorch-wavelets==1.2.3
PyYAML==5.3.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.3.3
sklearn==0.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
